VR Basics
You can move around with VR and in the direction you are looking by going forward.
Picking up objects is possible too in VR.
I want the VR player to be taller but having issues getting hands to raise, somewhere in the code they are being set to a position and I do not know where.
Going to work on making the player rotate by 45 degree angles with right joystick.